+++ src/mocks/mockILocale.ts
@@ -28,12 +28,15 @@ import keys from "lodash-es/keys";
 
 export class MockILocale {
     private currentLocale: string;
+    // eslint-disable-next-line @typescript-eslint/ban-types
     private locales: Object;
+    // eslint-disable-next-line @typescript-eslint/ban-types
     private static DefaultLocales: Object = {
         "en": "en-US",
         "ru": "ru-RU"
     };
 
+    // eslint-disable-next-line @typescript-eslint/ban-types
     constructor(locales: Object = MockILocale.DefaultLocales) {
         this.locales = locales;
         this.locale = keys(locales)[0];