+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.0
+* added `pointerEvent` method to test pointer events
+* moved to `eslint` from `tslint`
+* updated dependencies to fix vulnerabilities
+
+## 2.6.0
+* fix for `uuid`
 ## 2.5.0
 * powerbi-visual-api updated to 4.2.0
 * uuid package functionality replaced with crypto method